17_06_13_解决不同模拟器下int NSInteger的问题

+++ NZ Driving Test/NZ Driving Test/Classes/Home/Controller/YYHQuestionsCollectionViewController.m
@@ -141,7 +141,7 @@ static NSString * const reuseIdentifier = @"cell";
 
     YYHQuestionItem *questionItem = self.questionArray[indexPath.row];
     questionItem.current = indexPath;
-    questionItem.total = self.questionArray.count;
+    questionItem.total = (int)self.questionArray.count;
     cell.questionItem = questionItem;
 
     return cell;
